// src/app/core/enums/sector-id.enum.ts
/**
 * Enum de setores da empresa/mercado.
 */
export enum SectorIdEnum {
  INDUSTRIA = "industria",
  PETROQUIMICA = "petroquimica",
  LOGISTICA = "logistica",
  CONSTRUCAO = "construcao",
  COMBUSTIVEL = "combustivel",
  ALIMENTICIO = "alimenticio",
  AGRONEGOCIO = "agronegocio",
  HOSPITALAR = "hospitalar",
  ENERGIA = "energia",
  VAREJO = "varejo",
  QUIMICA = "quimica",
  METALURGIA = "metalurgia",
}